SPP family, ELAPFormerHead
