+++ mmseg/models/decode_heads/__init__.py
@@ -40,7 +40,8 @@ from .mlp_slowcatse_head import MLPSLowCatSEHead
 from .mlp_slowcatese_head import MLPSLowCatESEHead
 from .mlp_slowcat_la_head import MLPSLowCatLAHead
 from .ssformer_head import SSFormerHead
-from .lapformerpp_head import LAPFormerPPHead, LAPFormerPPHead_v2, LAPFormerPPHead_v3
+from .lapformerpp_head import LAPFormerPPHead, LAPFormerPPHead_v2, LAPFormerPPHead_v3, \
+    ELAPFormerHead
 
 __all__ = [
     'FCNHead', 'PSPHead', 'ASPPHead', 'PSAHead', 'NLHead', 'GCHead', 'CCHead',
@@ -55,5 +56,5 @@ __all__ = [
     'MLPSLowCatSEHead', 'MLPSLowCatESEHead', 'MLPSLowCatLAHead',
     'MLP_OSAHead_v2', 'SSFormerHead', 'MLP_OSAHead_v3', 'MLP_OSAHead_v4',
     'MLP_OSAHead_v5', 'MLP_OSAHead_v6', 'LAPFormerPPHead', 'LAPFormerPPHead_v2',
-    'LAPFormerPPHead_v3'
+    'LAPFormerPPHead_v3', 'ELAPFormerHead'
 ]
