+++ mmseg/models/decode_heads/lapformerpp_head.py
@@ -5,7 +5,7 @@ from mmcv.cnn import ConvModule
 from mmseg.models.builder import HEADS
 from mmseg.models.decode_heads.decode_head import BaseDecodeHead
 from mmseg.ops import resize
-from mmseg.models.utils import SELayer, CSPSPP, CSPSPPF, SPP, SPPF
+from mmseg.models.utils import SELayer, CSPSPP, CSPSPPF, SPP, SPPF, C3SPPF
 
 
 @HEADS.register_module()
@@ -327,6 +327,157 @@ class LAPFormerPPHead_v3(BaseDecodeHead):
             mode=self.interpolate_mode,
             align_corners=self.align_corners
         )
+
+        # perform identity mapping
+        out = outs[-1] + out
+
+        out = self.cls_seg(out)
+
+        return out
+
+
+@HEADS.register_module()
+class LAPFomerPPHead_v4(BaseDecodeHead):
+    def __init__(self,
+                 interpolate_mode='bilinear',
+                 pooling='CSPSPPF',
+                 **kwargs):
+        super(LAPFomerPPHead_v4, self).__init__(
+            input_transform='multiple_select',
+            **kwargs
+        )
+
+        assert pooling in ['SPSPP', 'CSPSPPF', 'SPP', 'SPPF', 'C3SPPF']
+        self.interpolate_mode = interpolate_mode
+
+        num_inputs = len(self.in_channels)
+        assert num_inputs == len(self.in_index)
+
+        self.convs = nn.ModuleList()
+        for i in range(num_inputs):
+            self.convs.append(
+                ConvModule(
+                    in_channels=self.in_channels[i],
+                    out_channels=self.channels,
+                    kernel_size=3,
+                    padding=1,
+                    norm_cfg=self.norm_cfg,
+                    act_cfg=self.act_cfg))
+
+        # feature fusion between adjacent levels
+        self.linear_projections = nn.ModuleList()
+        for i in range(num_inputs - 1):
+            self.linear_projections.append(
+                ConvModule(
+                    in_channels=self.channels * 2,
+                    out_channels=self.channels,
+                    kernel_size=1,
+                    stride=1,
+                    norm_cfg=self.norm_cfg,
+                    act_cfg=self.act_cfg
+                )
+            )
+
+        self.pooling = eval(pooling)(
+            in_channels=self.channels * num_inputs,
+            out_channels=self.channels * num_inputs
+        )
+
+
+@HEADS.register_module()
+class ELAPFormerHead(BaseDecodeHead):
+    def __init__(self,
+                 interpolate_mode='bilinear',
+                 pooling='SPPF'
+                 **kwargs):
+        super().__init__(input_transform='multiple_select', **kwargs)
+
+        self.interpolate_mode = interpolate_mode
+        assert pooling in ['SPSPP', 'CSPSPPF', 'SPP', 'SPPF', 'C3SPPF']
+        num_inputs = len(self.in_channels)
+
+        assert num_inputs == len(self.in_index)
+
+        # reduce channel dims and local emphasis
+        self.convs = nn.ModuleList()
+        for i in range(num_inputs):
+            self.convs.append(
+                ConvModule(
+                    in_channels=self.in_channels[i],
+                    out_channels=self.channels,
+                    kernel_size=3,
+                    padding=1,
+                    norm_cfg=self.norm_cfg,
+                    act_cfg=self.act_cfg))
+
+        # feature fusion between adjacent levels
+        self.linear_projections = nn.ModuleList()
+        for i in range(num_inputs - 1):
+            self.linear_projections.append(
+                ConvModule(
+                    in_channels=self.channels * 2 if self.ops == 'cat' else self.channels,
+                    out_channels=self.channels,
+                    kernel_size=1,
+                    stride=1,
+                    norm_cfg=self.norm_cfg,
+                    act_cfg=self.act_cfg
+                )
+            )
+
+        self.pooling = eval(pooling)(
+            in_channels=self.channels * num_inputs,
+            out_channels=self.channels * num_inputs
+        )
+
+        self.se_module = SELayer(
+            channels=self.channels * num_inputs
+        )
+        self.fusion_conv = ConvModule(
+            in_channels=self.channels * num_inputs,
+            out_channels=self.channels,
+            kernel_size=1,
+            norm_cfg=self.norm_cfg)
+
+    def forward(self, inputs):
+        # Receive 4 stage backbone feature map: 1/4, 1/8, 1/16, 1/32
+        inputs = self._transform_inputs(inputs)
+        _inputs = []
+        for idx in range(len(inputs)):
+            x = inputs[idx]
+            conv = self.convs[idx]
+            _inputs.append(
+                resize(
+                    input=conv(x),
+                    size=inputs[0].shape[2:],
+                    mode=self.interpolate_mode,
+                    align_corners=self.align_corners))
+
+        # slow concatenate
+        _out = torch.empty(
+            _inputs[0].shape
+        )
+        outs = [_inputs[-1]]
+        for idx in range(len(_inputs) - 1, 0, -1):
+            linear_prj = self.linear_projections[idx - 1]
+            # cat first 2 from _inputs
+            if idx == len(_inputs) - 1:
+                x1 = _inputs[idx]
+                x2 = _inputs[idx - 1]
+            # if not first 2 then cat from prev outs and _inputs
+            else:
+                x1 = _out
+                x2 = _inputs[idx - 1]
+            if self.ops == 'cat':
+                x = torch.cat([x1, x2], dim=1)
+            else:
+                x = x1 + x2
+            _out = linear_prj(x)
+            outs.append(_out)
+
+        out = torch.cat(outs, dim=1)
+        out = self.pooling(out)
+        out = self.se_module(out)
+        out = self.fusion_conv(out)
 
         # perform identity mapping
         out = outs[-1] + out
