+++ mmseg/models/decode_heads/mlp_osa_head.py
@@ -505,101 +505,3 @@ class MLP_OSAHead_v5(BaseDecodeHead):
         out = self.cls_seg(out)
 
         return out
-
-
-@HEADS.register_module()
-class MLP_OSAHead_v6(BaseDecodeHead):
-    def __init__(self,
-                 interpolate_mode='bilinear',
-                 ops='add',
-                 **kwargs):
-        super().__init__(input_transform='multiple_select', **kwargs)
-
-        self.interpolate_mode = interpolate_mode
-        assert ops in ['cat', 'add']
-        self.ops = ops
-        num_inputs = len(self.in_channels)
-
-        assert num_inputs == len(self.in_index)
-
-        # reduce channel dims and local emphasis
-        self.convs = nn.ModuleList()
-        for i in range(num_inputs):
-            self.convs.append(
-                ConvModule(
-                    in_channels=self.in_channels[i],
-                    out_channels=self.channels,
-                    kernel_size=3,
-                    padding=1,
-                    norm_cfg=self.norm_cfg,
-                    act_cfg=self.act_cfg))
-
-        # feature fusion between adjacent levels
-        self.linear_projections = nn.ModuleList()
-        for i in range(num_inputs - 1):
-            self.linear_projections.append(
-                ConvModule(
-                    in_channels=self.channels * 2 if self.ops == 'cat' else self.channels,
-                    out_channels=self.channels,
-                    kernel_size=3,
-                    padding=1,
-                    norm_cfg=self.norm_cfg,
-                    act_cfg=self.act_cfg
-                )
-            )
-
-        self.se_module = SELayer(
-            channels=self.channels * num_inputs
-        )
-        self.fusion_conv = ConvModule(
-            in_channels=self.channels * num_inputs,
-            out_channels=self.channels,
-            kernel_size=1,
-            norm_cfg=self.norm_cfg)
-
-    def forward(self, inputs):
-        # Receive 4 stage backbone feature map: 1/4, 1/8, 1/16, 1/32
-        inputs = self._transform_inputs(inputs)
-        _inputs = []
-        for idx in range(len(inputs)):
-            x = inputs[idx]
-            conv = self.convs[idx]
-            _inputs.append(
-                resize(
-                    input=conv(x),
-                    size=inputs[0].shape[2:],
-                    mode=self.interpolate_mode,
-                    align_corners=self.align_corners))
-
-        # slow concatenate
-        _out = torch.empty(
-            _inputs[0].shape
-        )
-        outs = [_inputs[-1]]
-        for idx in range(len(_inputs) - 1, 0, -1):
-            linear_prj = self.linear_projections[idx - 1]
-            # cat first 2 from _inputs
-            if idx == len(_inputs) - 1:
-                x1 = _inputs[idx]
-                x2 = _inputs[idx - 1]
-            # if not first 2 then cat from prev outs and _inputs
-            else:
-                x1 = _out
-                x2 = _inputs[idx - 1]
-            if self.ops == 'cat':
-                x = torch.cat([x1, x2], dim=1)
-            else:
-                x = x1 + x2
-            _out = linear_prj(x)
-            outs.append(_out)
-
-        out = torch.cat(outs, dim=1)
-        out = self.se_module(out)
-        out = self.fusion_conv(out)
-
-        # perform identity mapping
-        out = outs[-1] + out
-
-        out = self.cls_seg(out)
-
-        return out
