+++ mmseg/models/utils/__init__.py
@@ -12,12 +12,12 @@ from .visual_attention import VisualAttention, VisualAttentionv2
 from .weighted_VA import wVisualAttention
 from .layer_attention import LayerAttention, EfficientLayerAttn
 from .reversed_attention import ReversedAttention
-from .spp import SPP, CSPSPP, SPPF, CSPSPPF
+from .spp import SPP, CSPSPP, SPPF, CSPSPPF, C3SPPF
 
 __all__ = [
     'ResLayer', 'SelfAttentionBlock', 'make_divisible', 'InvertedResidual',
     'UpConvBlock', 'InvertedResidualV3', 'SELayer', 'PatchEmbed',
     'nchw_to_nlc', 'nlc_to_nchw', 'nchw2nlc2nchw', 'nlc2nchw2nlc',
     'VisualAttention', 'wVisualAttention', 'LayerAttention', 'EfficientSELayer',
-    'EfficientLayerAttn', 'GeSELayer', 'SPP', 'CSPSPP', 'SPPF', 'CSPSPPF'
+    'EfficientLayerAttn', 'GeSELayer', 'SPP', 'CSPSPP', 'SPPF', 'CSPSPPF', 'C3SPPF'
 ]
